requirements.txt
py-cord==2.6.1
python-dotenv==1.0.0
python-binance==1.0.19
pandas==2.1.4
